pyproject.toml
[tool.poetry]
name = "digital-market"
version = "0.1.0"
description = ""
authors = ["plamer <plamer88@yandex.ru>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10.6"
django = "^4.2.1"
djangorestframework = "^3.14.0"
pydantic = {extras = ["dotenv"], version = "^1.10.8"}
psycopg2-binary = "^2.9.6"
drf-spectacular = "^0.26.2"
djangorestframework-simplejwt = "^5.2.2"
drf-nested-routers = "^0.93.4"
django-filter = "^23.2"
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
